feat: allow to configure save per slots amount + abstract out indexer logic from `main`
